flaskProject/requirements.txt
Flask==1.1.2
tensorflow==2.4.0
Pillow==8.1.0
numpy==1.19.5